Update 7th jan ap.cpp

+++ 7th jan ap.cpp
@@ -4,11 +4,11 @@ public:
     
      int n = nums.size();
     int total_count = 0;
-    vector<unordered_map<long long, int>> dp(n);
+    vector<unordered_map<int, int>> dp(n);
 
     for (int i = 0; i < n; ++i) {
         for (int j = 0; j < i; ++j) {
-            long long diff = static_cast<long long>(nums[i]) - static_cast<long long>(nums[j]);
+            long long diff = (nums[i]) - (nums[j]);
             int count_j_diff = dp[j].count(diff) ? dp[j][diff] : 0;
             dp[i][diff] += count_j_diff + 1;
             total_count += count_j_diff;
